changed makefile and added audio read
